+++ doc/conf.py
@@ -18,6 +18,7 @@ extensions = [
     "sphinx.ext.viewcode",
     "sphinx_click",
     "sphinx_issues",
+    "sphinx_paramlinks",
     "sphinxcontrib.autodoc_pydantic",
     "myst_parser",
 ]