+++ scim2_tester/__init__.py
@@ -1,5 +1,7 @@
 from .checker import check_server
+from .utils import CheckConfig
 from .utils import CheckResult
+from .utils import SCIMTesterError
 from .utils import Status
 
-__all__ = ["check_server", "Status", "CheckResult"]
+__all__ = ["check_server", "Status", "CheckResult", "CheckConfig", "SCIMTesterError"]