+++ scim2_tester/checker.py
@@ -11,19 +11,21 @@ from scim2_tester.resource import check_resource_type
 from scim2_tester.resource_types import check_resource_types_endpoint
 from scim2_tester.schemas import check_schemas_endpoint
 from scim2_tester.service_provider_config import check_service_provider_config_endpoint
+from scim2_tester.utils import CheckConfig
 from scim2_tester.utils import CheckResult
 from scim2_tester.utils import Status
 from scim2_tester.utils import checker
 
 
 @checker
-def check_random_url(scim: BaseSCIMClient) -> tuple[Resource, CheckResult]:
+def check_random_url(conf: CheckConfig) -> tuple[Resource, CheckResult]:
     """Check that a request to a random URL returns a 404 Error object."""
     probably_invalid_url = f"/{str(uuid.uuid4())}"
-    response = scim.query(url=probably_invalid_url, raise_scim_errors=False)
+    response = conf.client.query(url=probably_invalid_url, raise_scim_errors=False)
 
     if not isinstance(response, Error):
         return CheckResult(
+            conf,
             status=Status.ERROR,
             reason=f"{probably_invalid_url} did not return an Error object",
             data=response,
@@ -31,49 +33,55 @@ def check_random_url(scim: BaseSCIMClient) -> tuple[Resource, CheckResult]:
 
     if response.status != 404:
         return CheckResult(
+            conf,
             status=Status.ERROR,
             reason=f"{probably_invalid_url} did return an object, but the status code is {response.status}",
             data=response,
         )
 
     return CheckResult(
+        conf,
         status=Status.SUCCESS,
         reason=f"{probably_invalid_url} correctly returned a 404 error",
         data=response,
     )
 
 
-def check_server(scim: BaseSCIMClient) -> list[CheckResult]:
+def check_server(scim: BaseSCIMClient, raise_exceptions=False) -> list[CheckResult]:
     """Perform a series of check to a SCIM server.
 
     It starts by retrieving the standard :class:`~scim2_models.ServiceProviderConfig`,
     :class:`~scim2_models.Schema` and :class:`~scim2_models.ResourceType` endpoints.
 
     Then for all discovered resources, it perform a series of creation, query, replacement and deletion.
+
+    :param scim: A SCIM client that will perform the requests.
+    :param raise_exceptions: Whether exceptions should be raised or stored in a :class:`~scim2_tester.CheckResult` object.
     """
+    conf = CheckConfig(scim, raise_exceptions)
     results = []
 
     # Get the initial basic objects
-    result_spc = check_service_provider_config_endpoint(scim)
+    result_spc = check_service_provider_config_endpoint(conf)
     service_provider_config = result_spc.data
     results.append(result_spc)
 
-    result_schemas = check_schemas_endpoint(scim)
+    result_schemas = check_schemas_endpoint(conf)
     results.append(result_schemas)
 
-    result_resource_types = check_resource_types_endpoint(scim)
+    result_resource_types = check_resource_types_endpoint(conf)
     resource_types = result_resource_types.data
     results.append(result_resource_types)
 
     # Miscelleaneous checks
-    result_random = check_random_url(scim)
+    result_random = check_random_url(conf)
     results.append(result_random)
 
     # Resource checks
     if result_resource_types.status == Status.SUCCESS:
         for resource_type in resource_types:
             results.extend(
-                check_resource_type(scim, resource_type, service_provider_config)
+                check_resource_type(conf, resource_type, service_provider_config)
             )
 
     return results