+++ scim2_tester/schemas.py
@@ -1,14 +1,14 @@
-from scim2_client import BaseSCIMClient
 from scim2_models import Resource
 from scim2_models import Schema
 
+from .utils import CheckConfig
 from .utils import CheckResult
 from .utils import Status
 from .utils import checker
 
 
 @checker
-def check_schemas_endpoint(scim: BaseSCIMClient) -> tuple[Resource, CheckResult]:
+def check_schemas_endpoint(conf: CheckConfig) -> tuple[Resource, CheckResult]:
     """As described in RFC7644 §4 <rfc7644#section-4>`, `/ResourceTypes` is a mandatory endpoint, and should only be accessible by GET.
 
     .. todo::
@@ -19,9 +19,10 @@ def check_schemas_endpoint(scim: BaseSCIMClient) -> tuple[Resource, CheckResult]
         - Check that a 403 response is returned if a filter is passed
         - Check that the 'ResourceType', 'ServiceProviderConfig' and 'Schema' schemas are provided.
     """
-    response = scim.query(Schema)
+    response = conf.client.query(Schema)
     available = ", ".join([f"'{resource.name}'" for resource in response.resources])
     return CheckResult(
+        conf,
         status=Status.SUCCESS,
         reason=f"Schemas available are: {available}",
         data=response.resources,